Exception support in fibers.
The problem before was that it was impossible to copy the exception off the
throwing stack.  The trick is ... don't copy it!  Perform a context switch from inside the catch block, saving a
pointer to the exception.  When you rethrow you can't throw it directly (because throw xxx; will make a copy
anyway), so instead either special case exception types to throw a full copy, or wrap it up in a nested
exception that has a reference to the original exception.  Unit tests for all of the above :).