Use parent() instead of FilterStream::
So yeah, maybe that is a better way to access it, because it ensures that (a) you skip any "blocking"
implementation in MutatingFilterStream, and (b) if you're calling a helper method, the helper method doesn't do
a virtual function call back to the filter stream.

Also, improve strictness of MutatingFilterStream to make everything except read and write not work by default
(you're still responsible for read and write yourself).