Fix handling of "complete" for Ragel Parsers
Write unit test to check for a ragel parser not dumping out early because it's on a final state, but there's
more data left.  Extend ragel parsers to make final() the old complete(), and complete() means that no more
data could possibly be consumed. For FSMs with unambiguous final states with no leaving transitions, those mean
the same thing.  For FSMs that could be in final(), but could still process more data (URI, currently),
complete() should just unconditionally return false, to greedily eat as much data as possible.  Handle
"inferred" completion (the parser didn't consume everything handed to it, but it's not in error).