Break HTTP Client circular references
The connection had a shared_ptr to the request, and the request to the connection.  Everything got cleaned up
fine in regular code paths, but once error conditions were introduced, leaks and deadlocks were an almost
surety.  Fix it by making the connection only have a regular pointer to the request (and assert that those
pointers are gone in the request's destructor).

Also break a circular reference between request/response stream on the request object.  For requestStream(), the
request object will still own the request stream, but the request stream will only have a regular pointer back
to the request object.  On request object destruction, the regular pointer is removed, and the request
stream is re-parented to a blank stream.  This means that if the request object destructs before the request stream, the request will abort, but the request stream will still be valid (it will just assert if you try to write to it, since it
is no longer associated with anything).  For responseStream(), we want to enable the use case of just returning
the response stream to a caller, and throwing away the request object (not useful in the request case, because
you'll want to check the response status).  To enable this, request object has a weak pointer to the response
stream, and the response stream has a shared pointer to the request object.  The downside to this is that if you
lose your own pointer to the response stream, the request object can't revive it, and the request must be
aborted.  In short, you can't use temporaries for the result of responseStream()... you have to save it to a
local variable if you want to call more than one method on it.

Change-Id: I0fd8181deec818110ff9d16179c04ccb9c465fa5
Reviewed-on: http://gerrit.dechocorp.com/1707
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>