size_t Stream::read(void *, size_t) helper method.
Default implementation on Stream will adopt the passed in buffer into a new Buffer, then make the regular read
call.  Then it intelligently decides if it can return immediately, copy into the passed in buffer, or make a new allocation
so that it can be safely copied into the passed in buffer.

More efficient implementations on MemoryStream and BufferedStream.

Change-Id: I5baaaa53913717876dd4dd65bde2e43b3c8d9644
Reviewed-on: http://gerrit.dechocorp.com/2010
Reviewed-by: Brian Palmer <brian@mozy.com>