Work on the concept of a RequestBroker and a ConnectionBroker
I'm tired of writing loops and try-catches to handle redirects and re-connects for every HTTP request.  So
abstract it out.  Instead of using a connection callback as the basis for communication, use a RequestBroker
(which in turn presumably has a ConnectionBroker).  BaseRequestBroker deals with re-establishing dead
connections.  RedirectRequestBroker deals with redirects.  Which leaves HTTP users to deal with... exactly what
they need to deal with, and nothing else.

Change-Id: Ief52200f4ecc8bf7cb5a47aa6afdfd2eb8c6918d
Reviewed-on: http://gerrit.dechocorp.com/2021
Reviewed-by: Brian Palmer <brian@mozy.com>