Fix assertion when pipelined http requests fail.
scheduleAllWaitingResponses() was removing it from m_pendingRequests (and needs to, to avoid race conditions in a multi-threaded environment), so just rearrange the checks for prior request failed to be before the assertion that we're where we think we are in m_pendingRequests.

Also simplify aborting a request, fixing a bug where waiting responses don't get scheduled, and fix a deadlock bug where the fiber could switch out while holding the mutex.

Change-Id: Ia82794ef3392864a55d380dafdc42c5097dc2767
Reviewed-on: http://gerrit.dechocorp.com/2605
Reviewed-by: Brian Palmer <brian@mozy.com>