Fix re-entrancy problem of ClientRequest::cancel() in the presence of exceptions
When the response stream is the only ref to a ClientRequest, and an exception occurs while reading from the stream, ClientRequest::cancel() would be called, and remove the response stream's reference.  This would immediately cause the destructor to be called, calling cancel (which would return ok), but then ASSERT because the connection still has pointer to us (because the original call to cancel() has not completed yet; in a release build the assertion would be gone, and would simply crash because this would have been deleted).  So if we have a response stream (and therefore a guaranteed reference), grab an extra reference for the duration of the cancel().

Change-Id: Ibf929f7d43d06963222a6879583cec1e93c56507
Reviewed-on: http://gerrit.dechocorp.com/2863
Reviewed-by: Cody Cutrer <cody@mozy.com>