Change how MORDOR_ASSERT works
First, always log it as a fatal error (which no longer terminates the process), then if a global is set to throw an exception, throw an exception like before.  Otherwise, terminate the process immediately instead of hoping that nobody else screws yup while unwinding from the exception, en route to terminating the process.  Set the setting in the unit tests, because they do expect to catch assertions.

Change-Id: I730af418f8be6af5de3c2cf823afcb8fb4b8a4fa
Reviewed-on: http://gerrit.dechocorp.com/3569
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>