Fix a crash with lots of pipelined requests
Remove the request from the connection if it's still there when resuming after waiting, but we have an error on the connection.  This can happen when it gets scheduled (so is no longer waiting to be scheduled) before a problem happens, and then the problem happens to clear other things out, but doesn't know to clear this request out.

Change-Id: If68072b2c277b9d8c23705b7ec7186b7c6965f54
Reviewed-on: http://gerrit.dechocorp.com/3571
Reviewed-by: Brian Palmer <brian@mozy.com>