Add optional parameter to Stream::flush() if it should flush the parent stream as well.
Default is true, previous behavior.  Adjust all filter streams to either pass it on directly (because they don't have anything that would make sense to flush), or handle it correctly.
Fix SSLStream unit test for fix of SSLStream not flushing parent when it should, and the unit test expects that it doesn't.

Change-Id: If63b19123226dd2f30481f0c5373a04017eae8ff
Reviewed-on: http://gerrit.dechocorp.com/3610
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>