Optimize IOManagerIOCP
No longer keep detailed state about pending IO operations.  Simply arrange the AsyncEventIOCP structure in such a way as to easily find the state when an event completes.  Removes the need for locking.  Keep that state around in debug builds so that if your IOManager won't shut down, you can have a prayer of finding the IO op that's blocking it.
Change all users of IOManagerIOCP to check the OVERLAPPED structure for their results (contrary to their names, Internal and InternalHigh are actually documented as the error code and the number of bytes transferred).
Fix bug in Socket::connect for Windows 2000 reporting an error when the connect succeeded.
Fix bug in GetQueuedCompletionStatusEx copying the Internal from the OVERLAPPED to the OVERLAPPED_ENTRY.  PostQueuedCompletionStatus makes so guarantees that lpOverlapped is actually an OVERLAPPED, so just don't trust it.  IOManagerIOCP doesn't even access that field anyway.

Change-Id: Iafd17fe51da30ff75ce3dc91d5db1d7c85501fec
Reviewed-on: http://gerrit.dechocorp.com/3665
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>