Remove support for inline-handshaking in SSLStream
It was too complicated keeping track of writes that could happen during reads, and could cause problems if a write blocked a read, but we ran out of buffer space somewhere, and the server was doing the same thing, it would deadlock.  So explicit connect() and accept() are required now, and attempting to re-negotiate from the other end will throw an exception (which isn't all that bad of an idea, anyway, due to the SSL re-negotiation vulnerability).
This simplified SSLStream quite a bit.
Also fixed that a close() could never really work because we'd send shutdown in close(), and wait for shutdown in flush(), but we'd already close()ed the parent stream.
Fix the SSLStream::duplexStress unit test to make sure that it really completed.

Change-Id: Ic2eee0c76ed452bf6c38a5645937c27717cd2114
Reviewed-on: https://gerrit.dechocorp.com/5502
Reviewed-by: Jeremy Stanley <jeremys@decho.com>