Fix for responseStream() being called after cancel() corrupting internal state
ensureResponse() was getting called again, but due to wonky logic, was resetting m_responseState to BODY, which then allowed cancel() to be called again, and we'd ASSERT inside of cancel trying to disconnect ourself from the connection.

So just simplify the logic in ensureResponse() to capture the exact exception that it previously threw (if any), and rethrow it.  If no exception, and we've already been called, simply return.

Change-Id: I1fb83caeccfe3d283b1b7b486d6464e09ea40eec
Reviewed-on: https://gerrit.dechocorp.com/5253
Reviewed-by: Jeremy Stanley <jeremys@decho.com>