Improve interface of Buffer and Stream
Buffer::readBuffer and writeBuffer now both include an a parameter to specify if you want it to coalesce buffers to meet the requested length, or instead just return as much as it can from a single buffer.
Improve the Stream "helper" functions so that each Stream only has to implement one of the two overloads for read and write.  Default emulation behavior for calling the Buffer overload when only the void * overload available is to *not* coalesce buffers.  There are additional protected overloads if you want to change this behavior (now used by SSLStream).
Change several streams to implement only the void * overload instead of the Buffer overload, as well as add the void * overload to some.
Change-Id: If0df9532f9b4e0acc625e6ae1aa3d49b07008fcc
Reviewed-on: https://gerrit.dechocorp.com/5885
Reviewed-by: Jeremy Stanley <jeremys@decho.com>