Avoid tickle storms, and fix stopping multithreaded schedulers.
There's the possibility that when stop is called and there is still work to be
done, the "stop" tickle will get lost.  So compensate by re-tickling for each
thread that does actually stop (i.e. the thread that completes the last piece
of work will tickle other threads that were waiting for work).

Change-Id: I71578e03f0f7ce871844a70d0fd2ada1f38b70d8
Reviewed-on: https://gerrit.dechocorp.com/9090
Reviewed-by: Jeremy Stanley <jeremys@decho.com>