Fix includes for not using a pch
And fix stuff so it compiles under Clang!!

Note, clang asplodes if you use most boost headers in a pre-compiled header, so
exclude them in the pch.

On my macbookpro, build time with gcc: 2:27, with clang: 3:12.
So a bit slower, but blame that on most of boost not being pre-compiled.

Change-Id: I769ab69728a8b7bc100050b966795b212f08eb9a
Reviewed-on: https://gerrit.dechocorp.com/9101
Reviewed-by: Cody Cutrer <cody@mozy.com>