Fix race condition with AsyncEvent and IOManagerIOCP
It was possible for AsyncEvent to be accessed by another thread (for example,
registering for another event) before getting NULLed out after scheduling it.
So, do the NULLing first, and don't touch it after scheduling the event.

The other IOManagers are not susceptible to this because they internally own
the AsyncEvent, and always lock it, whereas IOManagerIOCP works without locks,
with caller-provided AsyncEvents.

Change-Id: I88369d551312efa6576ed60e63cb3d88ac9163b4
Reviewed-on: https://gerrit.dechocorp.com/9539
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>