Fix idiosyncrasy with Fibers and exceptions
Basically, when an exception was caught by a fiber, the fiber
switched out in the middle of catch, and then we had to re-call the
fiber one more time to clean it up, and which point it would destruct
the exception on the stack, and really exit the fiber.  The destructor
of the exception could then access the current fiber, and things would
asplode because the current fiber is halfway destructed.

So, fix by not leaving the exception on the stack before we switch out,
and really exitting the fiber on an exception, instead of the confusing
half exit when an exception happens.

Change-Id: If888b6f16c93ced1e7566964592647ec08fcf993
Reviewed-on: https://gerrit.dechocorp.com/10304
Reviewed-by: Punky Brewster (labs admin) <punkyb@example.com>
Reviewed-by: Jeff Tulley <jefft@decho.com>
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>