Improve respondStream to support Range on unsizeable streams
And lots more test cases for it.  One behavior change is that it will
avoid doing a PARTIAL_CONTENT unless it is sure it can satisfy it
(i.e. *any* range request on an unsizeable stream must support
trailers, otherwise it will send the full entity).  This avoids
having to drop the connection if the range request couldn't be
satisfied after we've already told the client it could.  By using
trailers, we're able to see if we could satisfy the request first,
before telling the client which part of the file was returned.

The one slight inconsistency left is that a Range request such as
0-0 on an unsizeable stream that ends up being the whole stream
would have a status code of PARTIAL_CONTENT instead of OK.  I don't
think this should be a problem in practice.

Change-Id: I16da32525129a4c73eed2995ae6252ec51f48d96
Reviewed-on: https://gerrit.dechocorp.com/10405
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>