Socket::onRemoteClosed
The basic idea is to allow a notification that the other side of a
connection has been closed, even though there are no active operations
on the connection.  This enables a couple of scenarios.  First, it
allows HTTP::ConnectionCache to immediately drop the connection,
avoiding having to a do a retry since we currently don't find out
about the dropped connection until we try to write to it .
Second, it's great for a server application that is processing some
long running operation... it can cancel it if the client disappears.

Change-Id: I6035a121cf2f5dc9e4b79e120bd0bb7fa8d7a2b4
Reviewed-on: https://gerrit.dechocorp.com/12044
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>