Fix parallel_for with multi-threaded schedulers.
There are abundant race conditions with the same fiber being
scheduled multiple times on a multi-threaded scheduler.  So just bite
the bullet, use a mutex, and simplify the logic by moving most of it
into the impl fiber.

Also, remove support for returning false to break enumeration...
just throw an exception.

Change-Id: I2a2bf12a30b98e440122d33c0b9281ea7f4ad3eb
Reviewed-on: https://gerrit.dechocorp.com/12280
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>