add random IV option to CryptoStream
in this mode, a random initialization vector is created on
encryption and prepended to the ciphertext before encrypting.
it is split from the ciphertext on decryption.

also fix a bug where decryption fails when performing very small
reads, and expand unit tests to cover fixed IV, random IV, and
no IV (ECB) situations.

Change-Id: I6a1f8155165313d1eb4429c9e6e3df92915027b7
Reviewed-on: https://gerrit.dechocorp.com/16857
Reviewed-by: Cody Cutrer <cody@mozy.com>