Adding comments to Broker and HTTP interfaces
Mordor has a very useful HTTP library, but it tends to require
a complete knowledge of the code and HTTP spec/terminology to
make even a simple client request.  So far I had always just
cut and paste existing mordor code and then tried to adapt it
blindly but it was usually a recipe for frustation.

As part of my code jam time I've done a study of part of the
code and added more comments.  I tried to only document things
that I'm confident about, but please review, as a bad comment
can be more misleading than no comment at all.

http.h - Try to establish the relationship between the structures
there and the actual HTTP headers that the client and server
need to pass over the wire.

BTW it seems to be an oversight that there is no direct
Cookie support?  I added a comment with the workaround

uri.h - Maybe I'm the only dev unfamiliar with all the detailed
terminology for the parts of an URI, but I thought a few
examples would clarify what scheme, authority and fragment mean.
Also I found it is not clear how to properly build a query for an
URI, and there is a risk of accidental double encoding/decoding
if you don't ues it right.  So I put some specific tips on that
subject.

broker.h - RequestBrokers are a critical part of making a connection
but the abstraction and flexibility also makes it a bit obscure.
Another challenge is the fact that this header file exposes all sorts
of details that a normal user of the library wouldn't need to know.
This is a first pass at describing each class and some of the important
methods, with some emphasis on the ConnnectionCache.  (Note:
now that I understand it a bit better I think this code is very cool)

Change-Id: If90f1f571248d03392894b16e3208f01d3eebc92
Reviewed-on: https://gerrit.dechocorp.com/20605
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>
Reviewed-by: Dan Nuffer <dann@mozy.com>