fix deadlock in IOManagerEPoll
When one thread would get the event, it would schedule the fiber, and
another thread would run it.  The fiber would then immediately exit,
and when the triggering thread attempted to remove its reference to the
fiber, a destructor would fire (bound up in the fiber start routine)
that would then call cancelEvent on the same fd that triggered the event;
cancelEvent would then try to grab the mutex, and deadlock.

Unfortunately, I was not able to reproduce this with a unit test.  I
was only able to repro by manually inserting a sleep after scheduling,
but before releasing the fiber reference.