Fix crash when proxying with PAC scripts on Mac.
The call to evaluate a PAC script was crashing in the underlying
JavaScript VM:

JSC::Heap::markConservatively(JSC::MarkStack&, void*, void*)

The JS garbage collector does not work properly on a fiber-based thread,
so the solution was to isolate the call into a fiber-free thread. The
worker thread utilizes a message queue to communicate with the
requesting fiber-based thread. The fiber uses Mordor::sleep while
waiting to allow other fibers to continue.

Change-Id: If6a661bce5b320ed518f5183767ce8d7d1df31ef
Reviewed-on: https://gerrit.dechocorp.com/20915
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Andrew Skowronski <andrews@mozy.com>