Avoid auto proxy detection on each http request
The ProxyCache is typically used on every single http
request to look up the proxy rules for each url.
However autodetection has been found to take 15-30 seconds
to fail on some of our computers, which makes sync
completely unusable.

Autodetection is controlled in the Internet Options in Control
panel but the option is hard to find and seems to be on by default
on some machines.  So a fix in Mordor also seemed prudent.

Rather than always calling WinHttpGetProxyForUrl this fix now
calls the more explicit WinHttpDetectAutoProxyConfigUrl to perform
the autodetection.  If that fails it does not attempt again during the
lifetype of the ProxyCache, which for Kalypso and Sync is
the same as the lifetime of the RequestBroker.

The downside of this extra caching is the client needs to do
extra work to reset the cache if the network configuration changes.
e.g. if you move your laptop to a corporate network then an existing
RequestBroker would not pick up the need for a proxy unless
ProxyCache::resetDetectionResultCache is called.

fixes #790 (https://labs.dechocorp.com/issues/790)

Change-Id: I962e0fde6c5e54d812379662e38caf34b213684c
Reviewed-on: https://gerrit.dechocorp.com/21069
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>