Fix Mordor XML parser for attribute values
Parser was crashing if attribute value contained
references, e.g. &quot; or &apos;
For example a="&quot;" or b='&apos;' (but a="'" is legal xml)

The existing regular expression seemed to being trying
to call a separate callback for each reference, as is done
for the inner text.  I was not able to fix that and in any
case it would be problematic for a parser to handle the
same callback invoked for both building inner text and
attribute values without more context info.
Instead I demonstrate in the unit test
how the attribute value callback could take care of converting
any references itself.

I also added a bit more general unit testing.  I found that comments
are not supported, but fixing that is beyond my ragel skills so
i just leave a unit test as a reminder.

Change-Id: I7c5f735269eabf6b80a123c6f3e0ff746713e180
Change-Id: Iec17341a32858c3d86b78140bc2c20383ed5c00f
Reviewed-on: https://gerrit.dechocorp.com/21366
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>
Reviewed-by: Rodi Reich <rodir@mozy.com>