make preparedstatement support using text as result format
current format is BINARY only. However this may not suit every situation.
A typical scenario is this:
when calling a stored procedure which returns void, the result format has to be TEXT.
A fatal_error will be returned if the result format is BINARY.

There are a couple of stored procedures in demeter returning void. So this is kind of
a blocker to trogdor, which uses mordor to access demeter.

with this patch, BINARY is still the default format. While you can call
conn->prepare("...", "", PreparedStatement::TEXT) to specify the result format.

Signed-off-by: Dannis Song <danniss@mozy.com>
Change-Id: I22d49ef5e67ab946b82092c2146ff541b4c9576d
Reviewed-on: https://gerrit.dechocorp.com/21430
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Wilson <weiminp@mozy.com>