ClientRequest throw rather than assert
TritonClient was attempting to flush requestStream in its
error handling.  If the stream is in an error state then
just throw again, following similar logic to other methods
in ClientRequest.

Note: when the assert was ignored, e.g. in release build,
the eventual flush() call would end up throwing a
ConnectionAbortedException from the socket library.
So an alternative fix would be to adjust the assert and
let the error case proceed along its old code path.

Reference:
https://labs.dechocorp.com/issues/816

Change-Id: I75b69274828d535cc6012e4da4c9d985a3190b98
Reviewed-on: https://gerrit.dechocorp.com/21692
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Jeremy Stanley <jeremy@mozy.com>