Fixing warnings and compilation errors.
Most of the warnings are about shadowed functions (not overriding all versions of a parent class virtual function).

Compilation errors were failure to properly detect linux and not reporting that it is an unsupported platform, initialization of a non-const pointer from a pointer in a structure passed by const reference (the pointer is const as well).

Allowing --without-boost-program-options to be used without breaking at link time.  boost::program_options needs a compiled library and will conflict with use when -D_GLIBCXX_DEBUG is used as a compilation option.  Disabling the library (which has no other ill effects because it is only used in the samples) is the only reasonable workaround for now.

Change-Id: I34075faa23e9d604a35680ea3c004b62c1955817
Reviewed-on: https://gerrit.dechocorp.com/28492
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Dan Nuffer <dann@mozy.com>