Fix proxy bypass list parsing.
Fix how we parse the proxy and bypass lists. This fixes the case where the
bypass list entry has a wildcard in the middle or at the end of the hostname
(e.g. "www.*.example.com" or "10.17.*"). I've also changed the code so that
the automatic settings override the local settings; this is how IE behaves.
This change doesn't fix the unsafe string comparisons in this code (comparing
Unicode strings using strcmp is not safe); which I will address in a
subsequent change.

I've also cleaned up the "detection cache" code somewhat, by adding some
additional comments and logging to try and make it clear what the code is
trying to accomplish.

This change also updates the mordor Xcode project so that it is possible
to build and run the mordor unit tests from the GUI on mac; the new target
is named 'run_tests'.

Finally, I've added a number of tests for the proxyFromList function in
mordor/tests/http_proxy.cpp.

Change-Id: Id69008c1bb6970bf0ac8c36809f9e663eba075e7
Signed-off-by: Jason Carreiro <jcarreiro@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/33588
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Andrew Skowronski <andrews@mozy.com>