Enhance NamedPipe stream
Stash will use this Windows specific NamedPipe
wrapper (see https://gerrit.dechocorp.com/#change,37099
for how I put it in usage)

AFAIK So far this is only used in one echo example,
which is why it is a bit lacking in functionality.

-Expose ablity to set flags so we can
create a Message based pipe (required when using CallNamedPipe
on the Client Side)

-Add an API to close a client connection without
fully closing the NamedPipe.

-Reduce the duplicated code by having both constructors
share an init method.

-Reducing logging level for successful NamedPipe creation.

refs #83442

Change-Id: Ia0f599cd67b9e879d037902824783ee9a2ea59aa
Signed-off-by: Andrew Skowronski <andrews@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/37131
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Jeff Tulley <jefft@vmware.com>
Reviewed-by: Russ Simons <russs@mozy.com>