Use SharingViolation for ERROR_LOCK_VIOLATION
Fix only impacts Windows

When trying to upload outlook files we hit
a violation reading the file.  however because
the error code is 33 instead of 32 the client
does not use the sharing violation code. Instead
a vague "NativeException" is thrown.

I think the situations are so similar that the
same exception can be used for both.

From winerror.h

// MessageId: ERROR_SHARING_VIOLATION
// The process cannot access the file because it is being used by another
// process.

// MessageId: ERROR_LOCK_VIOLATION
// The process cannot access the file because another process has locked a
// portion of the file.
//

I did a search in kalypso and SharingViolation
does not have special handling, I don't
see any sign that this change would have
an impact in that client.

refs #86645

Change-Id: I8f6bc97a4ac5b3ba0cdf1890ead6e784dbda1378
Signed-off-by: Andrew Skowronski <andrews@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/42251
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Zhaoding Wu <wuz@vmware.com>
Reviewed-by: Russ Simons <russs@mozy.com>