Add start time to the stdout of the mordor test suite runner.
This timestamp is in seconds, simply because it is easier to
look at than smaller intervals.  This helps easily distinguish
whether a sigle test blocked, or whether tests are simply
slow (due to other factors such as machine load) when
tests timeout (as they have been doing on recent stash builds).

Include a configVar (true by default) to allow this timestamp
to be easily removed if it is not wanted.

Signed-off-by: John Bowers <jbowers@vmware.com>
Change-Id: Ibacbe590c164e860e71cd7613cb9fb7bcff636f2
Reviewed-on: https://gerrit.dechocorp.com/44339
Reviewed-by: Andrew Skowronski <askowronski@vmware.com>