Catch logging exception in Scheduler::run
Several Stash crash dumps show that the attempt to log
a message on an exception in Scheduler::run() can cause an
unhandled exception.  This is because logging attempts to
show the current fiber, calling Fiber::getThis(). It seems that
in some cases that is pointing to a Fiber object that has already
been deleted, or maybe with a race condition on another thread
that is calling Fiber::getThis() to allocate a new Fiber.

Unfortunately I don't know how to reproduce this or other info
about the root cause.  And I'm not certain whether this would just
delay the crash to another point in the code rather than really
improving reliability. But it might permit the code to proceed
to a place where a better crash dump would be produced.

refs #98680

Change-Id: I1b89fdc59cdc42ff158d232b94957151ebbd1c89
Signed-off-by: Andrew Skowronski <andrews@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/53647
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Kevin Cai <kevinc@mozy.com>
Reviewed-by: Russ Simons <russs@mozy.com>