Disable 64 to 32 bit warnings on Mac
Disable 20 warnings occuring in the xcode
compilation of mordor and mordorprotobuf target.

Practically all the warnings are coming from boost and
protocol buffer library code in places where long and int are interchanged.

If there is a better way to fix these please, or if they are really valid
then let me know of a better fix.

Example log with the errors:

http://build.dechocorp.com/job/sync-release/dist=seblddev-mac/3557/consoleFull

Signed-off-by: Andrew Skowronski <andrews@mozy.com>
Change-Id: I5b2adfcc22625e44264d0776ed6e2eee3995d1df
Signed-off-by: Andrew Skowronski <andrews@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/57503
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Kefu Chai <chaik@mozy.com>