Bandwidth throttling works with multiple connections per host
refs #111665

The change in throttle.cpp was changing max to min when we should
"Never sleep for longer than a tenth of a second"

getActiveConnections() returns the number of connections that are
active and have pending requests.

Testing done: tested bandwidth throttling with 1-3 connections per
host. The total bandwidth used was in line with the throttle.


Signed-off-by: russs <russs@mozy.com>
Change-Id: I41531810e84abd898a52e330134e06a07df8c3c0
Reviewed-on: https://gerrit.dechocorp.com/59056
Reviewed-by: Pancho Avila <panchoa@mozy.com>
Reviewed-by: Jenkins <hudsonbuild@mozy.com>