Add IOCP error threshold.
refs #111932
refs #112528

- Allow a number of IOCP errors to ignore when calling
PostQueuedCompletionStatus in the tickle() method (the default value is 0).

- If we reach the number of specified errors in the specified timeframe we
will stop ignoring the errors.

- The Sync product has experienced some "Insufficient system resources" errors
returned by PostQueuedCompletionStatus, possibly indicating the the IOCP
queue is full. There is a possibility that this is a rare error that we can
recover from by ignoring it and continuing on.

- Be careful when using this method as it is possible for the corresponding
GetQueuedCompletionStatusEx call in idle() to wait infinitely which could cause
deadlock if the PostQueuedCompletionStatus failure is ignored. (Sync makes use
of many Timers which should ensure that GetQueuedCompletionStatusEx will not
ever wait infinitely.)

Change-Id: I586ba09e92a6112645740bf5b2f56e63c2f6c5cb
Signed-off-by: Mat Bess<mathew.bess@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/59456
Reviewed-by: Andrew Skowronski <andrews@mozy.com>
Reviewed-by: Jenkins <hudsonbuild@mozy.com>