Minor fix for Windows iomanager logging
The logging of successful iomanager calls was highly misleading
because it shows the lasterror.  When the call succeeded the
error has nothing to do with the call being logged, it is just
whatever most recent error happened on the thread.

For example, with this fix the output will look like this:

GetQueuedCompletionStatusEx(000001D8, 997): 1, (1) (0, "The operation completed successfully.")

Instead of:

GetQueuedCompletionStatusEx(0000028C, 32): 1, (1) (258, "The wait operation timed out.")

Tip: the error string is printed via this operator:
std::ostream &operator <<(std::ostream &os, error_t error)

Also add a few more comments to explain the alternative
registerEvent() signature and relationship between registerFile
and registerEvent().

Change-Id: If82f0d65fbf3664042f83c9014ee885cc9fc804b
Signed-off-by: Andrew Skowronski <andrews@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/59827
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Mat Bess <mathew.bess@mozy.com>