Demonstrate problem with using FiberEvent in multiple thread scenario.
New test that shows the danger of using FiberEvent to try to
coordinate destruction.  The main conclusion is not to use
FiberEvent at all!  But for education purposes these test scenarios
are informative.

Note: earlier patch versions demontrate a temporary sleep
in the mordor code that forces the test to hit the race condition
more easily.  That is removed and the bad code is disabled so that this
can safely run.

refs #114320

Change-Id: Ibd741abef6b29b8d029698ab4ec65d1db88cebae
Signed-off-by: Andrew Skowronski <andrews@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/60812
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Kevin Cai <kevinc@mozy.com>