fixes #122624 - toUtf8 crashes on Mac with Chinese chars
Zhaoding pinpointed crashes to convertion utility when
a string made up entirely of 3-byte chinese characters was converted.

Kevin pointed out true nature of the bug in comments patch 1 - an extra
byte for the null termination must be allocated, otherwise CFStringGetCString
was failing and MORDOR_NOTREACHED macro hit.

A simple unit test is added to verify conversion of a specific
problematic string.

Change-Id: I4edc4f06ecb63a231ef6edb773a02998ac72a0c9
Signed-off-by: Zhaoding <zhaoding.wu@mozy.com>
Signed-off-by: Andrew Skowronski <andrews@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/68939
Reviewed-by: Jenkins <hudsonbuild@mozy.com>