refs #123553 - Tolerance for "mordor.tempdir" without trailing slash
On posix the mordor.tempdir config var only worked if it included
a trailing slash which made it error prone to use.

Windows already adds in slash if missing so this fix makes the
behavior consistent.

Also add support for TMPDIR env variable to make this code more
consistent with the mkstemp call.

Order is:
-mordor.tempdir if specified
-TMPDIR if specified
-otherwise /tmp

And we retain ability for the caller to override that behavior by
passing an absolute path as the temp file prefix.

Change-Id: I9944b5ea90cdd878a55685dcf87960c704dcb1a1
Signed-off-by: Andrew Skowronski <andrews@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/70413
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Xicheng Wang <Xicheng.Wang@mozy.com>