Create seperate vcxproj for Visual Studio 2013
-compiler choice is encoded into the vcxproj so add another copy of vcxproj
for Visual Studio 2013.

-Currently lzma and protocol buffers are disabled from that build,
to avoid unnecessary dependency on clients that don't use that.
In future will need to find a way to make it conditional or split it
out to separate optional library.

-Any new .cpp or .h will have to be added in both places.
Fortunately this is rare and the xml format is simple so hopefully this
won't be too large of a burden.

-The new vcxproj file references a shared property file to reference
expected version of boost, openssl etc, similar to gerrit/79804.
All paths are relative to %winclientlib% which is the standard.
Mordor clients could potentially override this by adding more poperties earlier in the INCLUDE or LIBRARY env variables.

Change-Id: Ibf0802d3cb22d368cc06c1fe82ab62ee327976e8
Signed-off-by: Andrew Skowronski <andrews@mozy.com>
Reviewed-on: https://gerrit.dechocorp.com/79851