Extended vs2013 support
Mordor already had initial support for vs2013 for parts needed by
Rufus and Kalypso.  This extends (and hopefully completes) the porting
of the full mordor library as needed by sync, lynx and other VBI
projects.

1) vs2013 version of mordorprotobuf

Create new copy of mordorprotobuf that has been upgraded to vs2012

This is consistent with other projects in mordor.
When all vs2010 or 2008 clients are obsolete we can remove the older
mordorprotobuf project files.

2) Restore protobuf to mordor.v12.vcxproj

Fix requires access to protoc in winclientlib, so 3rd party
prop files are also updated based on expected naming convention
(see gerrit/109239).

However lzma remains disabled - we aren't using that compression
in production so trying to maintain the 3rd party libraries for
each version of visual studio is unnecessary work.

Reference to winclientlib\lib removed because
that old approach of using a single 7z file is only for visual
studio 2008 and 2010

Change-Id: I421fa2facb185565714199d694b55d590f911afe
Signed-off-by: Andrew Skowronski <Andrew.Skowronski@mozy.com>
Reviewed-on: https://gerrit.mozy.lab.emc.com/109241
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Solon Chen <solon.chen@dell.com>
Reviewed-by: Andrew Skowronski <Andrew.Skowronski@emc.com>