Specify boost:: for a few calls to ref and bind
This change is to eliminate some ambiguous references that were
encountered when building with newer Ubuntu toolchain.

The original patch used std::, which worked on the new compiler.
However, on the build server compiler, bind and ref are not part
of std. I changed them to boost which should be supported.

Signed-off-by: Joseph Cline <Joseph.Cline@dell.com>
Change-Id: I15739b40e7734961a2124bb566de95b06ef23b49
Reviewed-on: https://gerrit.mozy.lab.emc.com/109906
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Ryan Bird <Ryan.Bird@emc.com>
Reviewed-by: Andrew Skowronski <Andrew.Skowronski@emc.com>