Fixing mordor for new versions of gcc.
C++14 calls flags all destructors noexcept(true), so attempting to
use throw in them is an error.  The bad alternative is to set
those which would use a throwing assert to be noexcept(false), but
that is infectious, so any code which could call a one of those
destructors form its destructor would also need to be noexcept(false).

Fixing a misleading indentation problem, which gcc6 treats as a warning
and will end up being an error.

Fixing the linkage of the examples.  They were missing boost libraries,
lzma, and openssl.  How this compiled dynamically before is unknown.

Signed-off-by: Kevin Harris <kevin.harris@emc.com>
Change-Id: Ia40d8624c15e2030feec95469e1db4f4399194a4
Reviewed-on: https://gerrit.mozy.lab.emc.com/111510
Reviewed-by: Jenkins <hudsonbuild@mozy.com>
Reviewed-by: Andrew Skowronski <Andrew.Skowronski@emc.com>
Reviewed-by: Ryan Bird <Ryan.Bird@emc.com>