CMake linux fixes.
libboost_locale requires libicu* on most linux platforms (this may change
when we have everything in thirdparty-linux).
Changing the build script so it will default to a parallel build, but
something like "build-cmake.sh debug -- -j1" can be used to build without
multiple cores.

Unrelated to the linux fix, but here for general use anyway is a new set of
functions for specifying libraries/binaries in a somewhat convenient syntax:
  add_binary(NAME foo SRCS foo.cpp TYPE EXE)
  add_exe(NAME bar SRCS bar.cpp LINUX_SRCS linux_specific.cpp LIBS xml)
  add_static_lib(NAME baz SRCS baz.cpp quux.cpp)

Yes, it is easier to read them when it is spread across multiple lines.


Signed-off-by: Kevin Harris <Kevin.Harris@mozy.com>
Change-Id: I2e77d3a3919cb66df364be64bbbd464d46b52eda
Reviewed-on: https://gerrit.mozy.lab.emc.com/114034
Reviewed-by: Andrew Skowronski <Andrew.Skowronski@emc.com>